Resource payments: model, controller, views and policies, all works

+++ app/controllers/users_controller.rb
@@ -1,19 +1,26 @@
 class UsersController < ApplicationController
-  #before_filter :authenticate_user!
-  #after_action :verify_authorized
+  before_filter :authenticate_user!
+  after_action :verify_authorized
 
   
   def index
     @users = User.all
-    #authorize User
+    authorize User
+  end
+
+  def show
+    @user = User.find(params[:id])
+    authorize @user
   end
 
   def new
     @user = User.new
+    authorize User
   end
 
   def create
     @user = User.new(secure_params)
+    authorize User
     if @user.save
       flash[:success] = "Your account has been created succesfully"
       session[:user_id] = @user.id
@@ -23,18 +30,14 @@ class UsersController < ApplicationController
     end
   end
 
-  def show
-    @user = User.find(params[:id])
-    #authorize @user
-  end
-
   def edit
     @user = User.find(params[:id])
+    authorize User
   end
 
   def update
     @user = User.find(params[:id])
-    #authorize @user
+    authorize @user
     if @user.update_attributes(secure_params)
       redirect_to users_path, :notice => "User updated."
     else
@@ -43,12 +46,13 @@ class UsersController < ApplicationController
   end
 
   def delete
-    @loan = Loan.find(params[:id])
+    @user = User.find(params[:id])
+    authorize @user
   end
 
   def destroy
     user = User.find(params[:id])
-    #authorize user
+    authorize user
     user.destroy
     redirect_to users_path, :notice => "User deleted."
   end
